fiin.py
This program simply takes a list of files on the command line, and lists the contents of each, with lines numbered.
There are a couple of new things from sys used here.

sys.argv
This gives the list of command line arguments. The first one (position zero) is the name of the script, as in C.
sys.exc_info
This returns a list of information about the current exception. The first item is the name of the exception, the second is the detail information, and the third is the call stack trace-back.