io.py
This program gives several examples of output formatting. One of the major tools in this area is the % operator. It is an operator which takes a string as its left argument, and a single value or a tuple as its right argument. It then examines the string (left argument) for % constructs of the sort which C printf uses, matches them in order with data items from the right, and forms a new string string substituting the data values for the % constructs.
Since the % returns the resulting string, it is most like sprintf in C. This string may then be printed with the usual Python print statement. There is one very important difference from C: The Python % operator is able to check that number and types of the data items match the number types of the % constructs, and throws an exception if these do not match. That's generally more polite than C, which cannot check, and will generally core dump or do something bizarre if these do not agree.